+++ modules/bridge/src/main/ts/ephox/bridge/core/ComponentSchema.ts
@@ -24,7 +24,7 @@ export const optionalSelect = FieldSchema.optionFunction('select');
 
 export const active = FieldSchema.defaultedBoolean('active', false);
 export const borderless = FieldSchema.defaultedBoolean('borderless', false);
-export const disabled = FieldSchema.defaultedBoolean('disabled', false);
+export const enabled = FieldSchema.defaultedBoolean('enabled', true);
 export const primary = FieldSchema.defaultedBoolean('primary', false);
 export const defaultedColumns = (num: number | 'auto'): FieldProcessor => FieldSchema.defaulted('columns', num);
 export const defaultedMeta = FieldSchema.defaulted('meta', {});