+++ modules/tinymce/src/plugins/media/test/ts/browser/ContentFormatsTest.ts
@@ -12,6 +12,7 @@ describe('browser.tinymce.plugins.media.ContentFormatsTest', () => {
     media_live_embeds: false,
     document_base_url: '/tinymce/tinymce/trunk/tests/',
     extended_valid_elements: 'script[src|type]',
+    allow_conditional_comments: true
   }, [ Plugin ]);
 
   it('TBA: Object retained as is', () => {