+++ modules/tinymce/src/themes/silver/main/ts/ui/button/MenuButton.ts
@@ -21,8 +21,8 @@ import { ToolbarButtonClasses } from '../toolbar/button/ButtonClasses';
 export type MenuButtonSpec = Omit<Toolbar.ToolbarMenuButton, 'type'>;
 
 const getMenuButtonApi = (component: AlloyComponent): Toolbar.ToolbarMenuButtonInstanceApi => ({
-  isDisabled: () => Disabling.isDisabled(component),
-  setDisabled: (state: boolean) => Disabling.set(component, state),
+  isEnabled: () => !Disabling.isDisabled(component),
+  setEnabled: (state: boolean) => Disabling.set(component, !state),
   setActive: (state: boolean) => {
     // Note: We can't use the toggling behaviour here, as the dropdown for the menu also relies on it.
     // As such, we'll need to do this manually