+++ modules/tinymce/src/themes/silver/test/ts/headless/components/selectbox/SelectboxWithSizeTest.ts
@@ -15,7 +15,7 @@ describe('headless.tinymce.themes.silver.components.selectbox.SelectboxWithSizeT
       name: 'selector',
       size: 5,
       label: Optional.some('selector'),
-      disabled: false,
+      enabled: true,
       items: [
         { value: 'one', text: 'One' },
         { value: 'two', text: 'Two' },